refactor: BookingFragment 파일 구조 변경 및 ReservationFragment로 네이밍 변경
