+++ README.md
@@ -168,3 +168,6 @@
 - [ ] 누르면 예매 정보 화면을 볼 수 있다.
 - [ ] 재실행해도 기존 예매 내역이 유지된다.
 - [ ] 예매 내역은 날짜와 시간 순으로 정렬된다.
+
+### 3단계 변경 사항
+- [x] refactor: BookingFragment 파일 구조 변경 및 ReservationFragment로 변경
