+++ app/src/main/java/woowacourse/movie/reservation/ReservationFragment.kt
@@ -1,18 +1,19 @@
-package woowacourse.movie
+package woowacourse.movie.reservation
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import woowacourse.movie.R
 
-class BookingFragment : Fragment() {
+class ReservationFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View {
-        val view = inflater.inflate(R.layout.fragment_booking, container, false)
+        val view = inflater.inflate(R.layout.fragment_reservation, container, false)
         return view
     }
 }
