+++ app/src/main/res/layout/reservation_item.xml
@@ -6,7 +6,7 @@
     android:layout_height="80dp">
 
     <TextView
-        android:id="@+id/textview_booking_date"
+        android:id="@+id/textview_reservation_date"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="12dp"
@@ -18,58 +18,58 @@
         tools:text="2024.4.1" />
 
     <TextView
-        android:id="@+id/booking_separator"
+        android:id="@+id/reservation_separator"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="2dp"
         android:layout_marginTop="12dp"
         android:text="@string/booked_separator"
         android:textSize="12sp"
-        app:layout_constraintStart_toEndOf="@+id/textview_booking_date"
+        app:layout_constraintStart_toEndOf="@+id/textview_reservation_date"
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/textview_booking_time"
+        android:id="@+id/textview_reservation_time"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="2dp"
         android:layout_marginTop="12dp"
         android:textSize="12sp"
-        app:layout_constraintStart_toEndOf="@+id/booking_separator"
+        app:layout_constraintStart_toEndOf="@+id/reservation_separator"
         app:layout_constraintTop_toTopOf="parent"
         tools:text="13:00" />
 
     <TextView
-        android:id="@+id/booking_separator2"
+        android:id="@+id/reservation_separator2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="2dp"
         android:layout_marginTop="12dp"
         android:text="@string/booked_separator"
         android:textSize="12sp"
-        app:layout_constraintStart_toEndOf="@+id/textview_booking_time"
+        app:layout_constraintStart_toEndOf="@+id/textview_reservation_time"
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/textview_booking_theater"
+        android:id="@+id/textview_reservation_theater"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="2dp"
         android:layout_marginTop="12dp"
         android:textSize="12sp"
-        app:layout_constraintStart_toEndOf="@+id/booking_separator2"
+        app:layout_constraintStart_toEndOf="@+id/reservation_separator2"
         app:layout_constraintTop_toTopOf="parent"
         tools:text="선릉 극장" />
 
     <TextView
-        android:id="@+id/textview_booking_movie_title"
+        android:id="@+id/textview_reservation_movie_title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="12dp"
         android:layout_marginTop="6dp"
         android:textSize="16sp"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textview_booking_date"
+        app:layout_constraintTop_toBottomOf="@+id/textview_reservation_date"
         tools:text="영화 제목" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
